refactor(extension): split background coordinators
