+++ packages/extension/scripts/test-store-artifact.mjs
@@ -61,12 +61,16 @@ try {
     "icons/icon-32.png",
     "icons/icon-48.png",
     "manifest.json",
+    "native_host_bridge.js",
     "options.css",
     "options.html",
     "options.js",
+    "overlay_coordinator.js",
     "popup.css",
     "popup.html",
     "popup.js",
+    "session_store.js",
+    "tab_group_manager.js",
   ].sort();
   const expectedLocaleContents = [
     "ar",
